change to module versus class design in pytest
