+++ python/pytest/test_bugs.py
@@ -1,0 +1,23 @@
+import torch
+import ptens
+from torch.autograd.gradcheck import gradcheck
+
+def test_bug1(device):
+    nnodes = 15
+    graph = ptens.ggraph.random(nnodes, 0.5)
+    subgraphs = [ptens.subgraph.trivial(), ptens.subgraph.edge()]
+    node_values = torch.rand(nnodes, 1, requires_grad=True)
+
+    node_attributes = ptens.subgraphlayer0.from_matrix(graph, ptens.subgraph.trivial(), node_values)
+
+    for sg in subgraphs:
+        gather_features = ptens.subgraphlayer0.gather(sg, node_attributes)
+        result = torch.sum(gather_features)
+        result.backward()
+
+        # linmap_features = ptens.subgraphlayer0.linmaps(node_attributes)
+        result = torch.sum(node_attributes)
+        result.backward()
+
+        check = gradcheck(ptens.subgraphlayer0.gather, (sg, node_attributes), eps=1e-3)
+        assert check
